requirements.txt
streamlit
ultralytics
opencv-python-headless
numpy<2.0.0
pillow
easyocr
python-doctr
nepali-datetime
torch==2.2.2
torchvision==0.17.2